+++ src/react/index.js
@@ -1,0 +1,4 @@
+export { useRAG } from './useRag.js';
+export { initRAG } from '../initRag.js';
+export { createBrowserModelClient } from '../browserModelClient.js';
+